ner/requirements.txt
# basic
pandas
numpy
# grpc
grpcio
grpcio-tools
# db
pymongo[srv]
dnspython
# testing
pytest